app/requirements.txt
flask
psycopg2-binary
prometheus-client